markeetplace_server/requirements.txt
Django
dj-rest-auth
drf-yasg
django-cors-headers
djangorestframework
django-allauth
pillow
django-environ
setuptools
django-admin-interface
stripe
django-ckeditor